Collapse the datetimes package into one module
Three files for one transformer and five small helpers made the reader
follow imports across a package to see how a date column is handled.
Everything now sits in `preprocessing/datetimes.py`, and the two test
files become one. The import path and behaviour are unchanged.
