--- tests/test_preprocessing/test_datetimes/__init__.py
@@ -1,1 +1,0 @@
-#  Copyright (c) Prior Labs GmbH 2026.
